backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.6.7
google-generativeai==0.7.2
python-dotenv==1.0.1
pymongo==4.6.1
